Listen to binlog connector lifecycle events
